Rename firmware updater to "firmware-updater" everywhere
