+++ firmware-update/metadata.yml.in
